#54 Fix "slidy g w componente -s"  #52 Validate existing file

+++ lib/src/modules/start.dart
@@ -11,24 +11,28 @@ import 'package:slidy/src/modules/install.dart';
 
 start(hasForce, completeStart) async {
   var dir = Directory("lib");
-  if (dir.listSync().isNotEmpty) {
-    if (hasForce) {
-      output.msg("Removing lib folder");
-      await dir.delete(recursive: true);
-    } else {
-      output.error("The lib folder must be empty");
-      exit(1);
+  if (await dir.exists()) {
+    if (dir.listSync().isNotEmpty) {
+      if (hasForce) {
+        output.msg("Removing lib folder");
+        await dir.delete(recursive: true);
+      } else {
+        output.error("The lib folder must be empty");
+        exit(1);
+      }
     }
   }
 
   var dirTest = Directory("test");
-  if (dirTest.listSync().isNotEmpty) {
-    if (hasForce) {
-      output.msg("Removing test folder");
-      await dirTest.delete(recursive: true);
-    } else {
-      output.error("The test folder must be empty");
-      exit(1);
+  if (await dirTest.exists()) {
+    if (dirTest.listSync().isNotEmpty) {
+      if (hasForce) {
+        output.msg("Removing test folder");
+        await dirTest.delete(recursive: true);
+      } else {
+        output.error("The test folder must be empty");
+        exit(1);
+      }
     }
   }
